requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-utils-workflow
llama-parse
python-dotenv
streamlit